message_encoder_watermark/message-watermarking/requirements.txt
flask
torch
torchvision
pillow
numpy
